Revert cost centers and tax groups platform changes (#1148)
* Revert "FYLE-2a81g3v: Moved tax group service to platform and added tax group models (#1107)"

This reverts commit 17f67ec401ef01ca2fef11756a43862006d555ca.

* Revert "FYLE-29ptkye-1: Added platform costcenter models and moved costcenter service methods to platform (#1115)"

This reverts commit 58dcf981ff8c50d7a26d4110266c70f27e7e74fe.
