--- src/app/core/models/platform/platform-api-response.model.ts
@@ -1,5 +1,0 @@
-export interface PlatformApiResponse<T> {
-  count: number;
-  offset: number;
-  data: T[];
-}
