--- src/app/core/models/platform/platform-cost-center.model.ts
@@ -1,10 +1,0 @@
-export interface PlatformCostCenter {
-  id: number;
-  name: string;
-  code: string;
-  description: string;
-  is_enabled: boolean;
-  org_id: string;
-  created_at: Date;
-  updated_at: Date;
-}
