--- src/app/core/models/platform/platform-tax-group.model.ts
@@ -1,10 +1,0 @@
-export interface PlatformTaxGroup {
-  id: string;
-  org_id: string;
-  created_at: Date;
-  updated_at: Date;
-  name: string;
-  percentage: number;
-  is_enabled: boolean;
-  code: string;
-}
