+++ src/app/core/services/cost-centers.service.ts
@@ -1,11 +1,7 @@
 import { Injectable } from '@angular/core';
+import { ApiService } from './api.service';
 import { Cacheable } from 'ts-cacheable';
-import { Observable, range, Subject } from 'rxjs';
-import { SpenderPlatformApiService } from './spender-platform-api.service';
-import { CostCenter } from '../models/v1/cost-center.model';
-import { concatMap, map, reduce, switchMap } from 'rxjs/operators';
-import { PlatformCostCenter } from '../models/platform/platform-cost-center.model';
-import { PlatformApiResponse } from '../models/platform/platform-api-response.model';
+import { Subject } from 'rxjs';
 
 const costCentersCacheBuster$ = new Subject<void>();
 
@@ -13,61 +9,17 @@ const costCentersCacheBuster$ = new Subject<void>();
   providedIn: 'root',
 })
 export class CostCentersService {
-  constructor(private spenderPlatformApiService: SpenderPlatformApiService) {}
+  constructor(private apiService: ApiService) {}
 
   @Cacheable({
     cacheBusterObserver: costCentersCacheBuster$,
   })
-  getAllActive(): Observable<CostCenter[]> {
-    return this.getActiveCostCentersCount().pipe(
-      switchMap((count) => {
-        count = count > 50 ? count / 50 : 1;
-        return range(0, count);
-      }),
-      concatMap((page) => this.getCostCenters({ offset: 50 * page, limit: 50 })),
-      map((res) => res),
-      reduce((acc, curr) => acc.concat(curr), [] as CostCenter[])
-    );
-  }
-
-  getActiveCostCentersCount(): Observable<number> {
+  getAllActive() {
     const data = {
       params: {
-        is_enabled: 'eq.' + true,
-        offset: 0,
-        limit: 1,
+        active_only: true,
       },
     };
-    return this.spenderPlatformApiService
-      .get<PlatformApiResponse<PlatformCostCenter>>('/cost_centers', data)
-      .pipe(map((res) => res.count));
-  }
-
-  getCostCenters(config: { offset: number; limit: number }): Observable<CostCenter[]> {
-    const data = {
-      params: {
-        is_enabled: 'eq.' + true,
-        offset: config.offset,
-        limit: config.limit,
-      },
-    };
-    return this.spenderPlatformApiService
-      .get<PlatformApiResponse<PlatformCostCenter>>('/cost_centers', data)
-      .pipe(map((res) => this.transformFrom(res.data)));
-  }
-
-  transformFrom(platformCostCenter: PlatformCostCenter[]): CostCenter[] {
-    const oldCostCenter = platformCostCenter.map((costCenter) => ({
-      active: costCenter.is_enabled,
-      code: costCenter.code,
-      created_at: costCenter.created_at,
-      description: costCenter.description,
-      id: costCenter.id,
-      name: costCenter.name,
-      org_id: costCenter.org_id,
-      updated_at: costCenter.updated_at,
-    }));
-
-    return oldCostCenter;
+    return this.apiService.get('/cost_centers', data);
   }
 }
