+++ src/app/core/services/tax_group.service.spec.ts
@@ -1,6 +1,6 @@
 import { TestBed } from '@angular/core/testing';
 
-import { TaxGroupService } from './tax-group.service';
+import { TaxGroupService } from './tax_group.service';
 
 xdescribe('TaxGroupService', () => {
   let service: TaxGroupService;
